i2c: bl808: use quirks to specify max len
Maximal transfer len should be specified using quirks as this is the
standard method. The i2c-core will make sure our master_xfer callback is
not called with messages that are too long.

It is important as some of the i2c clinet code (and regmap code) checks
this field to understand the adapter capabilities.