i2c: bl808: fix the order in case of TXF
In the current code, the bl808_fill_tx_fifo() is called first (if we
still have some data to send). It will decrease the msg_buf_remaining
field, possibility to 0. If we have more messages to send, the next if
statement will be true and a bl808_i2c_start_transfer() will be called
for the next message without waiting for actually finishing the message
that we have just filled the FIFO for.

This patch changes the order, to:
1. (unchanged) If no more bytes to send and no more messages, we're
   done.
2. Otherwise if we don't have bytes to send (but we have more messages
   because otherwise we would leave at step 1), prepare next message for
   sending.
3. Otherwise, we have some data to send in current message so fill the
   TX FIFO up.