i2c: bl808: fill TX FIFO before starting enabling the controler
In current code, the TX FIFO is only filled when we get the TXF
interrupt. This means that at the moment the transaction is started by
setting the M_EN bit in CONFIG register, the TX FIFO is empty.
This should trigger both the TXF (because TX FIFO is empty) and FER
(with TX underflow) because the controller has nothing to send. We don't
know what is the exact condition that trigger TX underflow because
apparently in most cases we get TXF first but maybe if we are not fast
enough, we may actually get FER?

Let's remove that uncertainty by filling the TX FIFO before starting the
transaction.