i2c: bl808: simplify the interrupt mask/unmask
The interrupt is always masked when it is disabled and it is unmasked
when it is enabled. For this reason there is no point in having both
functions.
This patch simplifies passing arguments to the interrupt management
functions by leveraging the fact that enable bits are shifted by
constant offset from enable bits. This not only shortens the code but
also removes unneeded additional read/write to the register.