i2c: bl808: fix multimessage case
Current code, in case of multiple messages would start sending the next
message in TXF interrupt. This is not correct as when we get TXF
interrupt, the fifo may not yet be drained by the controller. The new
message should be sent from END interrupt instead.

With this fix, the multi-message sending works. This can be checked
using something like this on SSD1306:

i2ctransfer -y 0 w16@0x3c 0x01 0x02 0x03 0x04 0x05 0x06 0x07 0x08 0x09
0x0a 0xb 0xc 0xd 0xe 0xf 0x10 w4@0x3c 0x21 0x22 0x23 0x24

Without this patch, the first message consisting of 16 bytes would be
cut somewhere around 0xc. With this patch, it is working fine.