i2c: bl808: do not write CLR bits to INT_STS
It was observed that sometimes the CR_I2C_END_CLR, CR_I2C_NAK_CLR and
CR_I2C_ARB_CLR bits are set when the INT_STS register is read. Those
bits are supposed to be W1C, so they should always read 0, but
apparently that is not always the case. Since writing ones there has a
side effect, it's best not do this.

This patch clears those 3 bits when INT_STS is written to enable/disable
the interrupts.

Signed-off-by: Krzysztof Adamski <k@japko.eu>